Stake in the ground: basic flow client
This is extremely rudimentary. There seem to be any number of corner-
cases around what dill wants from a binary stream (such as a .readline
call) and we were tripping over at least one issue with the completer's
multipart serialisation code.

Nonetheless, this is a step in the right direction.

Needs tests around much of this, plus the API fleshing out. (The plan
is to use decorators to make flow declarations work.)